Move EncodableKeyMacro and DecodableKeyMacro to CodableKeyMacro.swift
EncodableKeyMacro and DecodableKeyMacro implementations were moved from their own files into CodableKeyMacro.swift. The separate source files were deleted to consolidate macro definitions and simplify code organization.
